+++ packages/database/src/schemas/file.ts
@@ -19,7 +19,6 @@ import { FileSource } from '@/types/files';
 import { idGenerator } from '../utils/idGenerator';
 import { accessedAt, createdAt, timestamps } from './_helpers';
 import { asyncTasks } from './asyncTask';
-import { chunks } from './rag';
 import { users } from './user';
 
 export const globalFiles = pgTable('global_files', {
@@ -107,35 +106,6 @@ export const documents = pgTable(
 export type NewDocument = typeof documents.$inferInsert;
 export type DocumentItem = typeof documents.$inferSelect;
 export const insertDocumentSchema = createInsertSchema(documents);
-
-/**
- * 文档块表 - 将文档内容分割成块并关联到 chunks 表，用于向量检索
- * 注意：此表可选，如果已经使用 pages 字段存储了文档块，可以不需要此表
- */
-export const documentChunks = pgTable(
-  'document_chunks',
-  {
-    documentId: varchar('document_id', { length: 30 })
-      .references(() => documents.id, { onDelete: 'cascade' })
-      .notNull(),
-
-    chunkId: uuid('chunk_id')
-      .references(() => chunks.id, { onDelete: 'cascade' })
-      .notNull(),
-
-    pageIndex: integer('page_index'),
-
-    userId: text('user_id')
-      .references(() => users.id, { onDelete: 'cascade' })
-      .notNull(),
-
-    createdAt: createdAt(),
-  },
-  (t) => [primaryKey({ columns: [t.documentId, t.chunkId] })],
-);
-
-export type NewDocumentChunk = typeof documentChunks.$inferInsert;
-export type DocumentChunkItem = typeof documentChunks.$inferSelect;
 
 // @ts-ignore
 export const files = pgTable(
